requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Pillow==10.1.0
python-decouple==3.8
django-cors-headers==4.3.1
